+++ mpowering/utils/validate_urls.py
@@ -13,6 +13,8 @@ def run():
 
 
 if __name__ == "__main__":
+    import django
+    django.setup()
     run()