+++ docs/changelog.rst
@@ -1,2 +1,6 @@
 Change log
 =====================================
+
+
+
+* 26: How to add the actual files to the search index http://github.com/mPowering/django-mpowering/issues/issue/26